refactor: moved reservation dialog and service to separate module
